Update Pseudo Code
This needs to be altered to include curves and modified so that if there is a cross-over compared to just a turn the sumo-bot will either keep moving forward hit white, go back and follow the tight angle turn or make the tight angle turn but register black lines and go back to fix itself and continue following the right path. Also, there is no need to stop the timer. The timer just needs to reset when it needs to be started again. While statements are used for continuous paths. If statements are used if there is a change in the path which also changes the global variables.